organize generateInvestmentPlan in server
